Create manual submission update (#282)
Update create_manual_submission script

This script was written for the Hidden Treasures challenge. I've modified it to make it work for any challenge. The user now has to supply the submission name and description, and there can be any number of file inputs, which have to be of the form input_name=file.