PFDA-469 activejob vuln fix -upgrade (#337)
https://snyk.io/vuln/SNYK-RUBY-ACTIVEJOB-72640

active job needs to be updated to 4.2.11, that triggers rails upgrade to
4.2.11

text of the alert:

Overview
activejobdeclares job classes that can be run by a variety of queueing backends.

Affected versions of this package are vulnerable to Access Restriction Bypass. Carefully crafted user input can cause Active Job to deserialize it using GlobalId and allow an attacker to have access to information that they should not have.

Remediation
Upgrade activejob to version 4.2.11, 5.0.7.1, 5.1.6.1, 5.2.1.1 or higher.