Fix the session related code.
Since we upgraded rack to >=2.0.8, the way how we deal with the
session has been changed. In the new rack version the session id
was splitted in to two parts: a public id and a private id. We
need to store and operate with a secret session.private_id now.
Here is in more details: https://github.com/rack/rack/issues/1432#issuecomment-571688819

However, as it turned out, controller tests do not get session.id
set correctly (https://github.com/rspec/rspec-rails/issues/2255),

Still looking for a workaround.