PFDA-2673: GSRS substance should not be saved.
It turns out the GSRS database is supposed to read-only,
and that submitted substances are only for the purpose of
creating that JSON file in the user's home and does not need
to be inserted into the GSRS database.

Instead of passing the call through to the backend they want
us to return a success response after writing out the JSON file
so to trigger the correct response in the UI.

This does mean users won't see the newly submitted substance
after creating the JSON and in fact that is necessarily because
different groups using pFDA do not want others to see their
experimental substance before it is published.